Revert "Player can load game from a previous state by command line" Wrong branch
This reverts commit 0e21b2d36728485bf33852c9c989e0233ca38e9f.
